Actually closing the video in clip.py
Same pull request as the last one made by me, but this time for clip.py. MoviePy does not actually close the videofile, making it impossible to remove/move/make actions on it. This has been tested and works. If there are any other methods that are using the same MoviePy snippet, please edit with this fix I provided or let me know and I will make a pull request.